Move validation to it's own module, re-export validateAuthorizationArgs
